refactor: users entity 변경
