+++ src/main/java/com/doongji/nestalk/entity/user/Email.java
@@ -1,0 +1,5 @@
+package com.doongji.nestalk.entity.user;
+
+public class Email {
+
+}
